// apps/weight-loss-consultant-gateway/src/app/constants/enums.ts
export enum AccountStatus {
  ACTIVE = 1,
  INACTIVE = 0,
  PENDING = 2
}

export enum CampaignStatus {
  ACTIVE = 1,
  INACTIVE = 0,
  PENDING = 2
}

export enum PackageStatus {
  ACTIVE = 1,
  INACTIVE = 0,
  PENDING = 2
}

export enum Role {
  Customer = 'customer',
  Trainer = 'trainer',
  Admin = 'admin'
}